refactor: move editor files to src/editor/
