+++ src/editor/editor.tsx
@@ -1,5 +1,6 @@
 import { Editor as DraftJsEditor, EditorState as DraftJsEditorState } from 'draft-js';
 import * as React from 'react';
+import { prismDecorator } from './prismDecorator';
 
 type State = {
   editorState: DraftJsEditorState;
@@ -12,7 +13,7 @@ export class Editor extends React.Component {
     super(props);
 
     this.state = {
-      editorState: DraftJsEditorState.createEmpty(),
+      editorState: DraftJsEditorState.createEmpty(prismDecorator),
     };
 
     this.onChange = this.onChange.bind(this);
